WIP - move session mgmt and histroy to submod
